GetTestID
This API fetches the Test ID given the Test_code, which is helpful when a student or institute wants to know the test ID for the given test_code. Say a student wants to register for a test, then he enters the test_code and gets the test_id, which helps him register and load the test later.